Fixed bug with loading default state in standalone
There was a bug in the standalone plugin
where the default state, when loaded, would
try to set the filters when they didn't exist.

Fixing the bug (by checking for existing filters)
created a nother bug where the formantVisualiser would
not create an image. Set this to happen on prepareToPlayF function